Create CharacterListFactory
