+++ challangeMarvel/challangeMarvel/AppDelegate.swift
@@ -1,11 +1,3 @@
-//
-//  AppDelegate.swift
-//  challangeMarvel
-//
-//  Created by Larissa Diniz on 28/07/20.
-//  Copyright © 2020 Larissa Diniz. All rights reserved.
-//
-
 import UIKit
 
 @UIApplicationMain
@@ -15,13 +7,8 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
         window = UIWindow(frame: UIScreen.main.bounds)
-        let controller = CharactersListViewController()
-        let nav = UINavigationController(rootViewController: controller)
+        let nav = UINavigationController(rootViewController: CharacterListFactory.make())
         window?.rootViewController = nav
-        let worker = CharactersListWorker()
-        let presenter = CharactersListPresenter(view: controller)
-        let interactor = CharactersListInteractor(presenter: presenter, worker: worker)
-        controller.interactor = interactor
         window?.makeKeyAndVisible()
         return true
     }
